refactor: 다리 만들기 검증  기능 클래스 분리
