+++ src/test/java/bridge/BridgeMakerTest.java
@@ -1,0 +1,2 @@
+package bridge;public class BridgeMakerTest {
+}
